Add Grid class & update BSP class/tests
