+++ bot_1/lib/bot_saves_princess.rb
@@ -1,26 +1,13 @@
+require_relative 'grid'
+
 class BotSavesPrincess
   attr_reader :size, 
               :grid
 
-  def initialize(size, grid)
-    @size              = size
-    @grid              = grid
-    @bot_location      = locate_character_location("m")
-    @princess_location = locate_character_location("p")
-  end
-
-  # iterates over each row of the grid checking for the desired character
-  # returns two index array representing the vertical and horizontal location of character
-  def locate_character_location(character)
-    location = []
-
-    @grid.each_with_index do |element, index|
-      if element.include?(character)
-        location << index
-        location << (element.chars.find_index(character))
-      end
-    end
-    location
+  def initialize(grid)
+    @grid              = grid.grid
+    @bot_location      = grid.bot_location
+    @princess_location = grid.princess_location
   end
 
   # returns two index array representing the difference between the shared indexes
