+++ bot_1/spec/bot_saves_princess_spec.rb
@@ -1,15 +1,13 @@
+require './lib/grid'
 require './lib/bot_saves_princess'
 
 RSpec.describe BotSavesPrincess do 
-  context 'initialize' do 
-    let(:bsp) {BotSavesPrincess.new(3, ["---", "-m-", "p--"])}
+  context 'initialize' do
+    let(:grid) {Grid.new(3, ["---", "-m-", "p--"])}
+    let(:bsp) {BotSavesPrincess.new(grid)}
 
     it 'exists' do
       expect(bsp).to be_a(BotSavesPrincess)
-    end
-
-    it 'has a size attribute' do
-      expect(bsp.size).to eq(3)
     end
 
     it 'has a grid attribute' do
@@ -18,9 +16,12 @@ RSpec.describe BotSavesPrincess do
   end
 
   context 'determine_character_location_differences' do
-    let(:bsp_1) {BotSavesPrincess.new(3, ["---", "-m-", "p--"])}
-    let(:bsp_2) {BotSavesPrincess.new(3, ["--p", "-m-", "---"])}
-    let(:bsp_3) {BotSavesPrincess.new(5, ["-p---", "-----","--m--", "-----", "-----"])}
+    let(:grid_1) {Grid.new(3, ["---", "-m-", "p--"])}
+    let(:grid_2) {Grid.new(3, ["--p", "-m-", "---"])}
+    let(:grid_3) {Grid.new(5, ["-p---", "-----","--m--", "-----", "-----"])}
+    let(:bsp_1) {BotSavesPrincess.new(grid_1)}
+    let(:bsp_2) {BotSavesPrincess.new(grid_2)}
+    let(:bsp_3) {BotSavesPrincess.new(grid_3)}
 
     it 'can determine the differences between the character locations 3x3 grid' do
       expect(bsp_1.determine_character_location_differences).to eq([-1, 1])
@@ -33,9 +34,12 @@ RSpec.describe BotSavesPrincess do
   end
 
   context 'path to princess' do
-    let(:bsp_1) {BotSavesPrincess.new(3, ["---", "-m-", "p--"])}
-    let(:bsp_2) {BotSavesPrincess.new(3, ["--p", "-m-", "---"])}
-    let(:bsp_3) {BotSavesPrincess.new(5, ["-p---", "-----","--m--", "-----", "-----"])}
+    let(:grid_1) {Grid.new(3, ["---", "-m-", "p--"])}
+    let(:grid_2) {Grid.new(3, ["--p", "-m-", "---"])}
+    let(:grid_3) {Grid.new(5, ["-p---", "-----","--m--", "-----", "-----"])}
+    let(:bsp_1) {BotSavesPrincess.new(grid_1)}
+    let(:bsp_2) {BotSavesPrincess.new(grid_2)}
+    let(:bsp_3) {BotSavesPrincess.new(grid_3)}
 
     it 'can decipher path to princess 3x3 grid' do
       expect(bsp_1.display_path_to_princess).to eq(["DOWN", "LEFT"])
